refactor: create @packages/errors
