+++ package.json
@@ -181,7 +181,7 @@
     "snap-shot-it": "7.9.3",
     "start-server-and-test": "1.10.8",
     "stop-only": "3.0.1",
-    "strip-ansi": "4.0.0",
+    "strip-ansi": "6.0.0",
     "term-to-html": "1.2.0",
     "terminal-banner": "1.1.0",
     "through": "2.3.8",
