+++ packages/errors/README.md
@@ -1,0 +1,3 @@
+## @packages/errors
+
+Error definitions and utilities for Cypress
