+++ packages/errors/index.js
@@ -1,0 +1,5 @@
+if (process.env.CYPRESS_INTERNAL_ENV !== 'production') {
+  require('@packages/ts/register')
+}
+
+module.exports = require('./src')
