+++ packages/errors/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "extends": "../ts/tsconfig.json",
+  "include": [
+    "src"
+  ],
+  "exclude": [
+    "test",
+    "script"
+  ],
+  "compilerOptions": {
+    "strict": true,
+    "allowJs": false,
+    "rootDir": "src",
+    "outDir": "dist",
+    "noImplicitAny": true,
+    "resolveJsonModule": true,
+    "experimentalDecorators": true,
+    "noUncheckedIndexedAccess": true,
+    "importsNotUsedAsValues": "error",
+    "types": [
+      "node"
+    ]
+  }
+}
