+++ packages/server/lib/browsers/utils.ts
@@ -4,8 +4,7 @@ import type { FoundBrowser } from '@packages/launcher'
 import errors from '../errors'
 // @ts-ignore
 import plugins from '../plugins'
-
-const errorsChild = require('../errors-child')
+import { getError } from '@packages/errors'
 
 const path = require('path')
 const debug = require('debug')('cypress:server:browsers:utils')
@@ -140,7 +139,7 @@ function extendLaunchOptionsFromPlugins (launchOptions, pluginConfigResult, opti
   // interface and we need to warn them
   // TODO: remove this logic in >= v5.0.0
   if (pluginConfigResult[0]) {
-    options.onWarning(errorsChild.get(
+    options.onWarning(getError(
       'DEPRECATED_BEFORE_BROWSER_LAUNCH_ARGS',
     ))
