+++ packages/server/lib/plugins/child/browser_launch.js
@@ -1,5 +1,5 @@
 const util = require('../util')
-const errorsChild = require('../../errors-child')
+const { getError } = require('@packages/errors')
 
 const ARRAY_METHODS = ['concat', 'push', 'unshift', 'slice', 'pop', 'shift', 'slice', 'splice', 'filter', 'map', 'forEach', 'reduce', 'reverse', 'splice', 'includes']
 
@@ -21,7 +21,7 @@ module.exports = {
 
         hasEmittedWarning = true
 
-        const warning = errorsChild.get(
+        const warning = getError(
           'DEPRECATED_BEFORE_BROWSER_LAUNCH_ARGS',
         )
