+++ packages/server/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "./../ts/tsconfig.json",
   "include": [
     "lib/*.ts",
-    "lib/**/*.ts"
+    "lib/**/*.ts",
   ],
   "files": [
     "./../ts/index.d.ts"
